마이페이지 수정

+++ src/main/java/edu/kh/jvj/mypage/model/vo/Coupon.java
@@ -6,7 +6,7 @@ public class Coupon {
 	
 	private int couponNo; // 쿠폰 번호
 	private String couponName; // 쿠폰 이름
-	private int discountPer; // 할인률
+	private Double discountPer; // 할인률
 	
 	private Date provideDate; // 발급일
 	private Date expireDate; // 만료일
@@ -73,13 +73,13 @@ public class Coupon {
 
 
 
-	public int getDiscountPer() {
+	public Double getDiscountPer() {
 		return discountPer;
 	}
 
 
 
-	public void setDiscountPer(int discountPer) {
+	public void setDiscountPer(Double discountPer) {
 		this.discountPer = discountPer;
 	}
